requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib[bcrypt]
python-jose 